Extract pathway components and utils
Splits the large PathwayPanel.jsx into smaller, focused files:
- New components: BuildingList, LegendChip, ModificationSummary, PathwaySelect, SectionCard, TemplateSelect
- New utils.js for pure helpers (getTickStep, getErrorMessage, formatCompactTimestamp, resolveSelectedYear, getNodeFill, getNodeSize)
- STATUS_ACCENT moved to constants.js
- getTickStep moved from constants.js to utils.js
