+++ src/features/pathway/components/PathwayPanel.jsx
@@ -5,14 +5,12 @@ import {
   InputNumber,
   message,
   Modal,
-  Select,
   Spin,
   Typography,
 } from 'antd';
 import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 
 import InfoTooltip, { TooltipFromBackend } from 'components/InfoTooltip';
-import { ERROR_RED } from 'constants/theme';
 import { useInputs } from 'features/input-editor/hooks/queries/useInputs';
 import { useMapStore, COLOR_MODES } from 'features/map/stores/mapStore';
 import { getMainUseType } from 'features/map/utils/constructionColors';
@@ -20,7 +18,6 @@ import {
   BinAnimationIcon,
   CreateNewIcon,
   DuplicateIcon,
-  InputEditorIcon,
   RefreshIcon,
   RunIcon,
 } from 'assets/icons';
@@ -35,7 +32,21 @@ import {
 
 import 'features/project/components/Cards/OverviewCard/OverviewCard.css';
 import DuplicatePathwayModal from 'features/project/components/modals/DuplicatePathwayModal';
-import { STATUS_FILL, getTickStep } from '../constants';
+import BuildingList from './BuildingList';
+import LegendChip from './LegendChip';
+import ModificationSummary from './ModificationSummary';
+import PathwaySelect from './PathwaySelect';
+import SectionCard from './SectionCard';
+import TemplateSelect from './TemplateSelect';
+import { STATUS_ACCENT, STATUS_FILL } from '../constants';
+import {
+  formatCompactTimestamp,
+  getErrorMessage,
+  getNodeFill,
+  getNodeSize,
+  getTickStep,
+  resolveSelectedYear,
+} from '../utils';
 
 import {
   applyTemplatesToYear,
@@ -67,492 +78,6 @@ const MAX_VISIBLE_TIMELINE_LANES = 3;
 // Minimum horizon for the shared ruler so sparse, near-term pathways still render against a
 // long-term scale. Pathways that already run past this keep their own end year.
 const MIN_TIMELINE_END_YEAR = 2100;
-
-const STATUS_ACCENT = {
-  error: ERROR_RED,
-  changed: '#D97706',
-};
-
-const getErrorMessage = (error, fallbackMessage) => {
-  const detail = error?.response?.data?.detail;
-  if (typeof detail === 'string' && detail.trim()) {
-    return detail;
-  }
-  if (detail?.message) {
-    return detail.message;
-  }
-  if (error?.message) {
-    return error.message;
-  }
-  return fallbackMessage;
-};
-
-const formatCompactTimestamp = (value) => {
-  if (!value) {
-    return 'Not recorded';
-  }
-  const parsed = new Date(value);
-  if (Number.isNaN(parsed.getTime())) {
-    return String(value);
-  }
-  return parsed.toLocaleString([], {
-    dateStyle: 'medium',
-    timeStyle: 'short',
-  });
-};
-
-const pickClosestYear = (years, preferredYear) => {
-  if (!years?.length || preferredYear == null) {
-    return null;
-  }
-
-  return (
-    [...years].sort(
-      (left, right) =>
-        Math.abs(left - preferredYear) - Math.abs(right - preferredYear) ||
-        left - right,
-    )[0] ?? null
-  );
-};
-
-const resolveSelectedYear = ({
-  years,
-  preferredYear,
-  pendingYear,
-  currentYear,
-  rememberedYear,
-}) => {
-  const candidates = [preferredYear, pendingYear, currentYear, rememberedYear];
-
-  for (const candidate of candidates) {
-    if (candidate != null && years.includes(candidate)) {
-      return candidate;
-    }
-  }
-
-  const fallbackTarget =
-    preferredYear ?? pendingYear ?? currentYear ?? rememberedYear ?? null;
-  if (fallbackTarget != null) {
-    return pickClosestYear(years, fallbackTarget);
-  }
-
-  return years[0] ?? null;
-};
-
-const getNodeFill = (row) => {
-  const hasStale = row?.status?.has_stale_phase;
-  if (hasStale) return STATUS_ACCENT.error;
-  const primaryPhase = row?.status?.primary_phase ?? 'none';
-  return STATUS_FILL[primaryPhase] ?? STATUS_FILL.none;
-};
-
-const getNodeSize = () => 12;
-
-const LegendChip = ({ colour, label, outline, halo }) => (
-  <div style={{ display: 'inline-flex', alignItems: 'center', gap: 8 }}>
-    <span
-      style={{
-        width: 12,
-        height: 12,
-        borderRadius: 999,
-        display: 'inline-block',
-        background: colour,
-        border: outline
-          ? `2px solid ${outline}`
-          : '1px solid rgba(15, 23, 42, 0.08)',
-        boxShadow: halo ? `0 0 0 4px ${halo}` : 'none',
-      }}
-    />
-    <Text style={{ fontSize: 12, color: '#475569' }}>{label}</Text>
-  </div>
-);
-
-const SectionCard = ({ title, content, tooltipKey }) => (
-  <div
-    style={{
-      borderRadius: 14,
-      border: '1px solid rgba(148, 163, 184, 0.18)',
-      background: '#FFFFFF',
-      padding: 12,
-      minHeight: 60,
-    }}
-  >
-    <div
-      style={{
-        display: 'flex',
-        alignItems: 'center',
-        gap: 6,
-        paddingBottom: 8,
-        marginBottom: 8,
-        borderBottom: '1px solid #e0e0e0',
-      }}
-    >
-      <Text strong>{title}</Text>
-      {tooltipKey ? <InfoTooltip tooltipKey={tooltipKey} /> : null}
-    </div>
-    {content}
-  </div>
-);
-
-const BuildingPill = ({ name, color, onClick }) => (
-  <span
-    role={onClick ? 'button' : undefined}
-    tabIndex={onClick ? 0 : undefined}
-    onClick={onClick}
-    onKeyDown={
-      onClick
-        ? (e) => {
-            if (e.key === 'Enter') onClick();
-          }
-        : undefined
-    }
-    style={{
-      display: 'inline-block',
-      padding: '2px 8px',
-      borderRadius: 999,
-      fontSize: 11,
-      fontWeight: 500,
-      background: color ?? '#e8e8e8',
-      color: color ? '#fff' : '#475569',
-      cursor: onClick ? 'pointer' : 'default',
-    }}
-  >
-    {name}
-  </span>
-);
-
-const BuildingList = ({
-  buildings,
-  buildingColorMap,
-  rebuildCounts,
-  onBuildingClick,
-}) => {
-  if (!buildings?.length) return null;
-  return (
-    <div style={{ display: 'flex', flexWrap: 'wrap', gap: 4 }}>
-      {buildings.map((b) => {
-        const count = rebuildCounts?.[b] ?? 0;
-        const label = count > 0 ? `${b}(${count})` : b;
-        return (
-          <BuildingPill
-            key={b}
-            name={label}
-            color={buildingColorMap?.[b]}
-            onClick={onBuildingClick ? () => onBuildingClick(b) : undefined}
-          />
-        );
-      })}
-    </div>
-  );
-};
-
-const ModificationSummary = ({ row, constructionColorMap: colorMap }) => {
-  const modifications = row?.modifications ?? {};
-  const archetypes = Object.keys(modifications);
-  if (!archetypes.length) return null;
-  return (
-    <div style={{ display: 'flex', flexDirection: 'column', gap: 6 }}>
-      {archetypes.map((archetype) => {
-        const components = modifications[archetype] ?? {};
-        const color = colorMap?.[archetype];
-        const changes = Object.entries(components).flatMap(
-          ([component, fields]) =>
-            Object.entries(fields).map(([field, value]) => ({
-              component,
-              field,
-              value,
-            })),
-        );
-        return (
-          <div key={archetype} style={{ fontSize: 12 }}>
-            <BuildingPill name={archetype} color={color} />
-            {changes.map((c) => (
-              <div
-                key={`${c.component}-${c.field}`}
-                style={{ color: '#475569', paddingLeft: 8 }}
-              >
-                {c.field}: {String(c.value)}
-              </div>
-            ))}
-          </div>
-        );
-      })}
-    </div>
-  );
-};
-
-const PathwayOptionWithCheckbox = ({
-  pathwayName,
-  checked,
-  onToggle,
-  onDelete,
-  onDuplicate,
-}) => {
-  const [isHovered, setIsHovered] = useState(false);
-
-  return (
-    <div
-      style={{
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-      }}
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-    >
-      <div
-        style={{
-          display: 'flex',
-          alignItems: 'center',
-          gap: 8,
-          minWidth: 0,
-          flex: 1,
-        }}
-      >
-        <input
-          type="checkbox"
-          checked={checked}
-          onChange={(e) => {
-            e.stopPropagation();
-            onToggle(pathwayName);
-          }}
-          onClick={(e) => e.stopPropagation()}
-          style={{ cursor: 'pointer', flexShrink: 0 }}
-        />
-        <div
-          style={{
-            minWidth: 0,
-            overflow: 'hidden',
-            textOverflow: 'ellipsis',
-            whiteSpace: 'nowrap',
-          }}
-          title={pathwayName}
-        >
-          {pathwayName}
-        </div>
-      </div>
-      {isHovered && (
-        <div style={{ display: 'flex', alignItems: 'center', flexShrink: 0 }}>
-          <DuplicateIcon
-            style={{ padding: '2px 4px', cursor: 'pointer', opacity: 0.55 }}
-            onClick={(e) => {
-              e.stopPropagation();
-              onDuplicate?.(pathwayName);
-            }}
-          />
-          <BinAnimationIcon
-            style={{ padding: '2px 4px' }}
-            className="cea-job-info-icon danger shake"
-            onClick={(e) => {
-              e.stopPropagation();
-              onDelete?.(pathwayName);
-            }}
-          />
-        </div>
-      )}
-    </div>
-  );
-};
-
-const PathwaySelect = ({
-  selectedPathway,
-  visiblePathways,
-  overviewPathways,
-  onToggleVisible,
-  onDeletePathway,
-  onDuplicatePathway,
-  onCreatePathway,
-  loading,
-  allBaked,
-}) => {
-  const [open, setOpen] = useState(false);
-
-  const sortedPathways = useMemo(() => {
-    return [...overviewPathways]
-      .map((p) => p.pathway_name)
-      .sort((a, b) => a.toLowerCase().localeCompare(b.toLowerCase()));
-  }, [overviewPathways]);
-
-  const visibleSet = useMemo(() => new Set(visiblePathways), [visiblePathways]);
-
-  const options = useMemo(() => {
-    return sortedPathways.map((pathwayName) => ({
-      label: (
-        <PathwayOptionWithCheckbox
-          pathwayName={pathwayName}
-          checked={visibleSet.has(pathwayName)}
-          onToggle={onToggleVisible}
-          onDelete={onDeletePathway}
-          onDuplicate={onDuplicatePathway}
-        />
-      ),
-      value: pathwayName,
-    }));
-  }, [
-    sortedPathways,
-    visibleSet,
-    onToggleVisible,
-    onDeletePathway,
-    onDuplicatePathway,
-  ]);
-
-  const hasPathways = overviewPathways.length > 0;
-
-  const displayLabel =
-    visiblePathways.length > 0
-      ? visiblePathways.join('; ')
-      : (selectedPathway ?? '');
-
-  return (
-    <Select
-      className={`${visiblePathways.length === 1 && allBaked ? 'cea-scenario-select-blue' : 'cea-scenario-select'} ${!hasPathways || !selectedPathway || visiblePathways.length === 0 ? 'cea-select-empty cea-select-glow' : ''}`}
-      style={{ width: 208 }}
-      styles={{ popup: { root: { width: 270 } } }}
-      placeholder={hasPathways ? 'Select Pathway' : 'Create Pathway'}
-      options={hasPathways ? options : []}
-      value={visiblePathways.length > 0 ? selectedPathway : undefined}
-      onChange={() => {}}
-      onSelect={(pathwayName) => {
-        onToggleVisible(pathwayName);
-        setOpen(true);
-      }}
-      loading={loading}
-      open={hasPathways ? open : false}
-      onOpenChange={hasPathways ? setOpen : undefined}
-      onClick={!hasPathways ? onCreatePathway : undefined}
-      notFoundContent={<small>No pathways</small>}
-      labelRender={() =>
-        visiblePathways.length > 0 ? (
-          <span
-            style={{
-              overflow: 'hidden',
-              textOverflow: 'ellipsis',
-              whiteSpace: 'nowrap',
-            }}
-          >
-            {displayLabel}
-          </span>
-        ) : null
-      }
-    />
-  );
-};
-
-const TemplateOption = ({ templateName, description, onEdit, onDelete }) => {
-  const [isHovered, setIsHovered] = useState(false);
-
-  const handleEditClick = (e) => {
-    e.stopPropagation();
-    onEdit?.(templateName);
-  };
-
-  const handleDeleteClick = (e) => {
-    e.stopPropagation();
-    onDelete?.(templateName);
-  };
-
-  return (
-    <div
-      style={{
-        display: 'flex',
-        justifyContent: 'space-between',
-        alignItems: 'center',
-      }}
-      onMouseEnter={() => setIsHovered(true)}
-      onMouseLeave={() => setIsHovered(false)}
-    >
-      <div
-        style={{
-          minWidth: 0,
-          overflow: 'hidden',
-          textOverflow: 'ellipsis',
-          whiteSpace: 'nowrap',
-          flexGrow: 1,
-        }}
-        title={description ? `${templateName}: ${description}` : templateName}
-      >
-        {templateName}
-        {description ? (
-          <span style={{ color: 'rgba(0, 0, 0, 0.45)' }}>: {description}</span>
-        ) : null}
-      </div>
-      {isHovered && (
-        <div style={{ display: 'flex', alignItems: 'center', flexShrink: 0 }}>
-          <InputEditorIcon
-            style={{ padding: '2px 6px' }}
-            className="cea-job-info-icon"
-            title={`Edit '${templateName}'`}
-            onClick={handleEditClick}
-          />
-          <BinAnimationIcon
-            style={{ padding: '2px 8px' }}
-            className="cea-job-info-icon danger shake"
-            title={`Delete '${templateName}'`}
-            onClick={handleDeleteClick}
-          />
-        </div>
-      )}
-    </div>
-  );
-};
-
-const TemplateSelect = ({
-  templates,
-  descriptions,
-  selectedTemplates,
-  onSelectTemplates,
-  onEditTemplate,
-  onDeleteTemplate,
-  onCreateTemplate,
-  loading,
-}) => {
-  const [open, setOpen] = useState(false);
-
-  const sortedTemplates = useMemo(() => {
-    return [...(templates ?? [])].sort((a, b) =>
-      a.toLowerCase().localeCompare(b.toLowerCase()),
-    );
-  }, [templates]);
-
-  const options = useMemo(() => {
-    return sortedTemplates.map((name) => ({
-      label: (
-        <TemplateOption
-          templateName={name}
-          description={descriptions?.[name]}
-          onEdit={onEditTemplate}
-          onDelete={onDeleteTemplate}
-        />
-      ),
-      value: name,
-    }));
-  }, [sortedTemplates, descriptions, onEditTemplate, onDeleteTemplate]);
-
-  const hasTemplates = sortedTemplates.length > 0;
-
-  return (
-    <Select
-      mode="multiple"
-      optionLabelProp="value"
-      className={`cea-template-select ${!hasTemplates ? 'cea-select-empty cea-select-glow' : ''}`}
-      style={{ width: 416 }}
-      styles={{ popup: { root: { width: 416 } } }}
-      placeholder={
-        hasTemplates ? 'Intervention Templates' : 'Create Intervention Template'
-      }
-      options={hasTemplates ? options : []}
-      value={selectedTemplates}
-      onChange={onSelectTemplates}
-      maxTagCount="responsive"
-      allowClear={hasTemplates}
-      loading={loading}
-      open={hasTemplates ? open : false}
-      onOpenChange={hasTemplates ? setOpen : undefined}
-      onClick={!hasTemplates ? onCreateTemplate : undefined}
-      notFoundContent={<small>No intervention templates</small>}
-    />
-  );
-};
 
 const PathwayPanel = ({
   open,
