+++ src/features/pathway/constants.js
@@ -1,6 +1,6 @@
-// Shared constants and pure helpers for the pathway feature.
-// Imported by PathwayPanel, OverviewCard, BuildingLifecycleCard, and any other
-// components that render pathway timelines.
+// Shared constants for the pathway feature.
+
+import { ERROR_RED } from 'constants/theme';
 
 export const STATUS_FILL = {
   none: '#CBD5E1',
@@ -10,16 +10,7 @@ export const STATUS_FILL = {
   simulated: '#000000',
 };
 
-/**
- * Returns the smallest label step (in years) that keeps tick labels at least
- * 56 px apart at the given pixels-per-year scale.
- * @param {number} pxPerYear
- * @returns {number}
- */
-export const getTickStep = (pxPerYear) => {
-  const steps = [1, 2, 5, 10, 20, 50, 100, 200, 500];
-  for (const step of steps) {
-    if (pxPerYear * step >= 56) return step;
-  }
-  return 1000;
+export const STATUS_ACCENT = {
+  error: ERROR_RED,
+  changed: '#D97706',
 };
